add CRUD subcommands for Profile plugin
